requirements.txt
urllib3==1.26.6
google-cloud-firestore
openai
streamlit_chat
gtts
SpeechRecognition
streamlit_option_menu
streamlit_extras
st_pages